Refactored rolls to shorten code and limit complexity
moved all apology message check to begining to return apology if citeria is met without trying rest of the code. Added list comprehenstion to rolls to make the code more concise and easier to read. Skipped refactor for Call of Cthulu roll - I dont entirely get the rules and intention so left it as is. I dont want to mess this up :)